class khodro ezafe shod
