+++ src/main/java/khodro/Khodro.java
@@ -1,0 +1,15 @@
+package khodro;
+
+public class Khodro {
+    boolean ayaRoshanAst;
+    String name = "honda";
+
+    public void roshanKardan() {
+        ayaRoshanAst = true;
+    }
+
+    public void khamoshKardan() {
+        ayaRoshanAst = false;
+    }
+
+}
